// src/interfaces/AllMovies.tsx
// Generated by https://quicktype.io

export interface ApiMovies {
    data: Movies[];
    msg:  string;
}

export interface Movies {
    id:           number;
    title:        string;
    rating:       string;
    awards:       number;
    release_date: string;
    length:       number | null;
    genre_id:     number | null;
    createdAt:    null | string;
    updatedAt:    null | string;
}