// src/interfaces/OneMovie.tsx
// Generated by https://quicktype.io

import { Movies } from "./AllMovies";

export interface MovieDetail {
    data: Movie;
    msg:  string;
}

export interface Movie extends Movies {
  /*   id:           number;
    title:        string;
    rating:       string;
    awards:       number;
    release_date: string;
    length:       number;
    genre_id:     number;
    createdAt:    null;
    updatedAt:    null; */
    genre:        Genre;
    actors:       Actor[];
}

export interface Actor {
    id:                number;
    first_name:        string;
    last_name:         string;
    rating:            string;
    favorite_movie_id: number | null;
    createdAt:         null;
    updatedAt:         null;
    actor_movie:       ActorMovie;
}

export interface ActorMovie {
    actorId:   number;
    movieId:   number;
    createdAt: null;
    updatedAt: null;
}

export interface Genre {
    id:        number;
    name:      string;
    ranking:   number;
    active:    number;
    createdAt: string;
    updatedAt: null;
}

export interface DataFormMovie {
    title: string;
    rating: string;
    awards: string;
    release_date: string;
    length: string;
    genre_id: string;
}