// src/interfaces/OneGenre.tsx
// Generated by https://quicktype.io

import { Movies } from "./AllMovies";

export interface DetailGenre {
    data: GeneroDetail;
    msg:  string;
}

export interface GeneroDetail {
    id:        number;
    name:      string;
    ranking:   number;
    active:    number;
    createdAt: string;
    updatedAt: null;
    movies:    Movies[];
}

/* export interface Movie {
    id:           number;
    title:        string;
    rating:       string;
    awards:       number;
    release_date: string;
    length:       number;
    genre_id:     number;
    createdAt:    null;
    updatedAt:    null;
} */