// src/interfaces/AllActors.tsx
// Generated by https://quicktype.io

export interface TodosLosActores {
    data: Actors[];
    msg:  string;
}

export interface Actors {
    id:                number;
    first_name:        string;
    last_name:         string;
    rating:            string;
    favorite_movie_id: number | null;
    createdAt:         null;
    updatedAt:         null;
}